Support packaging with pypi. #33,#24,#12
