+++ requirements.txt
@@ -1,22 +1,32 @@
 atomicwrites==1.3.0
 attrs==19.1.0
+bleach==3.1.0
 certifi==2019.3.9
 chardet==3.0.4
 Click==7.0
 Deprecated==1.2.5
+docutils==0.14
 idna==2.8
 importlib-metadata==0.17
 more-itertools==7.0.0
 packaging==19.0
+pkginfo==1.5.0.1
 pluggy==0.12.0
 py==1.8.0
 PyGithub==1.43.7
+Pygments==2.4.2
 PyJWT==1.7.1
 pyparsing==2.4.0
 pytest==4.6.2
+readme-renderer==24.0
 requests==2.22.0
+requests-toolbelt==0.9.1
 six==1.12.0
+that-is-me-on-github==0.0.2
+tqdm==4.32.1
+twine==1.13.0
 urllib3==1.25.3
 wcwidth==0.1.7
+webencodings==0.5.1
 wrapt==1.11.1
 zipp==0.5.1
