--- that_is_me_on_github.md
@@ -1,86 +1,0 @@
-# That Is Me On Github
-## User Info
-username: [Lifei Chen](https://github.com/hustclf)
-
-email: [hustclf@gmail.com](hustclf@gmail.com)
-
-followers: 11
-
-![avatar](https://avatars1.githubusercontent.com/u/13963737?v=4 "avatar")
-
-        
-## Owned Repostories
-Total: 7
-        
-* [weixinCrawler](https://github.com/hustclf/weixinCrawler), stars: 9, folks: 7
-            
-* [that_is_me_on_github](https://github.com/hustclf/that_is_me_on_github), stars: 6, folks: 3
-            
-* [RateLimiter](https://github.com/hustclf/RateLimiter), stars: 6, folks: 0
-            
-* [PHPRedisSentinel](https://github.com/hustclf/PHPRedisSentinel), stars: 2, folks: 0
-            
-* [kafka-streams-examples](https://github.com/hustclf/kafka-streams-examples), stars: 1, folks: 0
-            
-* [log_generator](https://github.com/hustclf/log_generator), stars: 1, folks: 0
-            
-* [hustclf.github.io](https://github.com/hustclf/hustclf.github.io), stars: 0, folks: 0
-            
-## Pull Requests
-Total: 13
-
-RateLimiter
-            
-* [MINOR: update readme](https://github.com/hustclf/RateLimiter/pull/1) \[closed\] \[merged\]
-                
-kafka
-            
-* [KAFKA-8187 Bugfix for branch 2.0](https://github.com/apache/kafka/pull/6849) \[open\] \[not_merged\]
-                
-* [KAFKA-8187: Add wait time for other thread in the same jvm to free the locks](https://github.com/apache/kafka/pull/6818) \[closed\] \[merged\]
-                
-* [KAFKA-8348 document optimized](https://github.com/apache/kafka/pull/6707) \[closed\] \[merged\]
-                
-* [MINOR:Replace duplicated code with common function in utils](https://github.com/apache/kafka/pull/6819) \[closed\] \[merged\]
-                
-* [MINOR: Document improvement for JMH-Benchmark module](https://github.com/apache/kafka/pull/6682) \[closed\] \[merged\]
-                
-faust
-            
-* [MINOR: fix bug of example in doc](https://github.com/robinhood/faust/pull/353) \[closed\] \[merged\]
-                
-flink-deployer
-            
-* [Replace Cancel with Terminate](https://github.com/ing-bank/flink-deployer/pull/37) \[closed\] \[merged\]
-                
-* [Support new command: terminate](https://github.com/ing-bank/flink-deployer/pull/36) \[closed\] \[not_merged\]
-                
-* [Support new command: terminate](https://github.com/ing-bank/flink-deployer/pull/35) \[closed\] \[not_merged\]
-                
-kubeedge
-            
-* [doc: optimize slack url in readme and readme_zn](https://github.com/kubeedge/kubeedge/pull/255) \[closed\] \[merged\]
-                
-flink
-            
-* [[FLINK-11819][Documentation] Fix wrong description about attribute order by supported now](https://github.com/apache/flink/pull/7901) \[closed\] \[not_merged\]
-                
-kafkacat
-            
-* [Fix bugs for wrong format when consuming message with json format](https://github.com/edenhill/kafkacat/pull/166) \[closed\] \[merged\]
-                
-## Issues:
-Total: 4
-        
-flink-deployer
-            
-* [Questions about local development](https://github.com/ing-bank/flink-deployer/issues/33) \[closed\]
-                
-* [Any plan for supporting other commands?](https://github.com/ing-bank/flink-deployer/issues/26) \[closed\]
-                
-kafkacat
-            
-* [Consume message with json format is invalid](https://github.com/edenhill/kafkacat/issues/167) \[closed\]
-                
-* [kafkacat consumer with json format does not contain timestamp](https://github.com/edenhill/kafkacat/issues/165) \[closed\]
-
