+++ that_is_me_on_github/config.py
@@ -1,0 +1,1 @@
+VERSION = "v0.0.2"
